+++ src/main/java/com/meli/service/backend/mutant/logic/ResultsStatisticsLogic.java
@@ -1,10 +1,8 @@
 package com.meli.service.backend.mutant.logic;
 
-import java.math.BigDecimal;
-
 public interface ResultsStatisticsLogic {
 
     Integer amountOfMutantOrHuman(Boolean value);
-    BigDecimal averageMutants(Integer countMutants, Integer countHumans);
+    Float averageMutants(Integer countMutants, Integer countHumans);
 
 }
